refactor: rename embeddings to documents
